+++ CHANGELOG.md
@@ -137,6 +137,7 @@
   ([#5968](https://github.com/facebook/jest/pull/5968))
 * `[jest-config]` Ensure that custom resolvers are used when resolving the
   configuration ([#5976](https://github.com/facebook/jest/pull/5976))
+* `[pretty-format]` Handle React.forwardRef
 * `[website]` Fix website docs
   ([#5853](https://github.com/facebook/jest/pull/5853))